+++ test/test_generate.rb
@@ -1,0 +1,5 @@
+class TestGenerate; end
+
+class PostTest; end
+
+class CommentTest; end