+++ test/test_route.rb
@@ -1,0 +1,7 @@
+require 'svrless'
+
+class TestRoute
+  include ResourceGenerator
+
+  resources :post, :comment_resource, :user
+end