+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## [Unreleased]
 
+## [0.2.0]- 2021-02-14
+
+- Template generation is cleaner now.
+- Replace Event with ResourceGenerator class.
+- Moved template formation from ResourceGenerator into various generator classes.
+- Added tests to cover these.
+
+## [0.1.1] - 2021-02-14
+
+- Minor improvements to initial set up
+
 ## [0.1.0] - 2021-02-11
 
 - Initial release